refactor: consolidate the process into Runner
